refactor: phase 7 - move scanner routes to scanners blueprint
