+++ src/app.py
@@ -91,79 +91,6 @@ def clear_history():
         print(f"Error clearing history: {e}")
         return jsonify({'status': 'error', 'message': str(e)}), 500
 
-@app.route('/subdomain', methods=['GET', 'POST'])
-def subdomain_page():
-    subdomains = []
-    message = ""
-    deep_scan = False
-    
-    if request.method == "POST":
-        domain = request.form.get("domain", "").strip()
-        deep_scan = request.form.get("deep_scan") == "on"
-        
-        if domain:
-            try:
-                # Use the new deep subdomain scanner
-                results = scan_subdomains_blocking(domain, deep_scan=deep_scan)
-                
-                if results:
-                    # Format results for display
-                    formatted_results = []
-                    for result in results:
-                        if isinstance(result, dict):
-                            formatted_results.append(result)
-                        else:
-                            formatted_results.append({
-                                'subdomain': str(result),
-                                'status_code': None,
-                                'status_text': 'Found',
-                                'dns_records': {}
-                            })
-                    
-                    subdomains = formatted_results
-                    message = f"✅ Found {len(subdomains)} subdomain(s)"
-                else:
-                    message = "❌ No subdomains detected"
-                    
-            except Exception as e:
-                logger.error(f"Subdomain scan error: {e}")
-                message = "❌ Scan error: " + str(e)
-    
-    return render_template('subdomain.html', subdomains=subdomains, message=message, active_page='subdomain')
-
-@app.route('/directory', methods=['GET', 'POST'])
-def directory_page():
-    """Directory finder page"""
-    directories = []
-    message = ""
-    deep_scan = False
-
-    if request.method == "POST":
-        target = request.form.get("target", "").strip()
-        deep_scan = request.form.get("deep_scan") == "on"
-
-        if target:
-            try:
-                results = scan_directories_blocking(target, deep_scan=deep_scan)
-                if results:
-                    directories = results
-                    message = f"✅ Found {len(directories)} path(s)"
-                else:
-                    message = "❌ No directories detected"
-            except Exception as e:
-                logger.error(f"Directory scan error: {e}")
-                message = "❌ Scan error: " + str(e)
-
-    return render_template('directory.html', directories=directories, message=message, active_page='directory')
-
-@app.route('/database-vulnerability', methods=['GET', 'POST'])
-def database_vulnerability_page():
-    """Database vulnerability scanner page"""
-    vulnerabilities = []
-    message = ""
-    
-    return render_template('database_vulnerability.html', vulnerabilities=vulnerabilities, message=message, active_page='database-vulnerability')
-
 @app.route('/export/<scan_id>', methods=['GET'])
 def export_report(scan_id):
     history = load_history()
@@ -884,6 +811,8 @@ from routes.views import views_bp
 app.register_blueprint(views_bp)
 from routes.api import api_bp
 app.register_blueprint(api_bp)
+from routes.scanners import scanners_bp
+app.register_blueprint(scanners_bp)
 
 if __name__ == '__main__':
     PORT = int(os.environ.get('PORT', 5000))
@@ -896,4 +825,3 @@ if __name__ == '__main__':
     print("="*60)
     print("Press CTRL+C to stop the server\n")
     socketio.run(app, host=HOST, port=PORT, debug=True, allow_unsafe_werkzeug=True)
-
